Move discord mocks to dedicated directory
